honeypot/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration
google-generativeai==0.3.2
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Validation (optional, with fallbacks)
validators==0.22.0
tldextract==5.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0